+++ NEWS.md
@@ -8,6 +8,7 @@
 * mgr-applications FAR mode semver dependency validation fails with SNAPSHOT version ranges(MGRAPPS-68)
 * Fix test for tenant header Kong route expression updated in application-poc-tools (APPPOCTOOL-64)
 * FAR mode OOM crash due to excessive memory consumption during validation (MGRAPPS-67)
+* Consume entitlement Kafka events for Kong route mgmt (MGRAPPS-70)
 ---
 
 ## Version `v3.0.0` (12.03.2025)